recyclerView funcionando, RVList_recentactivities_Activity.kt criar alterações depois
