+++ app/src/main/res/layout/activity_rvlist_recentactivities.xml
@@ -8,7 +8,7 @@
     tools:context=".RVList_recentactivities_Activity">
 
     <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/rv_lista"
+        android:id="@+id/tv_title"
         android:layout_width="0dp"
         android:layout_height="0dp"
         app:layout_constraintBottom_toBottomOf="parent"
